Initial fork of the ashuffle integration container
ashuffle uses a special docker container for its test environment.
Linters, unit-tests, and integration tests, all need a complete
development environment to build and test the code. Since this requires
a large number of build packages, the build/test container takes a
looong time to build (2-5 minutes on travis-ci).

Instead of building the container on each test-run, we will build a
"base image" of the container once, and then re-use it (by pulling it
from the docker hub) each time we need to run a test or build ashuffle.

Hopefully this should allow the integration tests to run in under 20m.